Split refactor faq (#2592)
* Update FAQ to reflect Token Launch

This PR pushes a new content section from lines 559 - 627 that addresses questions related to the token launch. I have tried format everything as good as I possibly know/can, apologies if it needs to be reworked.

Questions are:
562 - Does CowSwap Have a token?
580 - Is there a proposal to make vCOW transferrable?
585 - what is the purspose of the COW tokens?
605 - what was the criteria for the token airdrop?

* Fix code style issues with Prettier

* Fix building error

* Add space

* Update index.tsx

* Fix code style issues with Prettier

* Split and refactor FAQ

* Fix links

* Add footer

* Add footer to index too

* Update metrics

* Fix internal links

* Hide back in home

* Link to FAQ

* Correct year

* Fix typo. Fixes #2067
