+++ src/custom/constants/index.ts
@@ -114,7 +114,7 @@ export const GAS_FEE_ENDPOINTS = {
   [ChainId.XDAI]: 'https://blockscout.com/xdai/mainnet/api/v1/gas-price-oracle',
 }
 
-export const UNSUPPORTED_TOKENS_FAQ_URL = '/faq#what-token-pairs-does-cowswap-allow-to-trade'
+export const UNSUPPORTED_TOKENS_FAQ_URL = '/faq/trading#what-token-pairs-does-cowswap-allow-to-trade'
 
 // fee threshold - should be greater than percentage, show warning
 export const FEE_SIZE_THRESHOLD = new Fraction(10, 100) // 30%
