+++ src/custom/pages/Faq/styled.tsx
@@ -1,0 +1,86 @@
+import { ExternalLink as ExternalLinkTheme } from 'theme'
+import styled from 'styled-components/macro'
+import { Content } from 'components/Page'
+
+export const ExternalLinkFaq = styled(ExternalLinkTheme)`
+  color: ${({ theme }) => theme.text1};
+  text-decoration: underline;
+  font-weight: normal;
+  transition: color 0.2s ease-in-out;
+
+  &:hover {
+    color: ${({ theme }) => theme.textLink};
+  }
+`
+
+export const Wrapper = styled.div`
+  #table-container {
+    margin: auto;
+    max-width: 100%;
+
+    > table {
+      width: 100%;
+      border-spacing: 1px;
+      color: ${({ theme }) => theme.text1};
+
+      > thead {
+        background: ${({ theme }) => theme.tableHeadBG};
+      }
+
+      > tbody > tr {
+        background: ${({ theme }) => theme.tableRowBG};
+      }
+
+      > tbody > tr > td > span[role='img'] {
+        font-size: 18px;
+      }
+
+      th,
+      td {
+        text-align: left;
+        padding: 6px 12px;
+
+        &:not(:first-child) {
+          text-align: center;
+        }
+      }
+
+      th {
+        padding: 16px 12px;
+      }
+    }
+  }
+
+  h2 {
+    color: ${({ theme }) => theme.primary1};
+  }
+
+  > div:not(:first-child) {
+    margin: 2rem 0;
+  }
+
+  ${Content} {
+    > div > ul {
+      margin: 12px 0 24px;
+      padding: 0 0 0 20px;
+      color: ${({ theme }) => theme.primary1};
+      line-height: 1.2;
+    }
+
+    > div > ul > li {
+      margin: 0 0 12px;
+    }
+
+    > h3 {
+      margin: 0;
+
+      ::before {
+        border-top: none;
+      }
+    }
+  }
+
+  ol > li {
+    margin-bottom: 0.5rem;
+  }
+`
