+++ src/custom/pages/Swap/SwapMod.tsx
@@ -1002,7 +1002,7 @@ export default function Swap({
             <>
               <p>CowSwap requires offline signatures, which is currently not supported by some wallets.</p>
               <p>
-                Read more in the <HashLink to="/faq#wallet-not-supported">FAQ</HashLink>.
+                Read more in the <HashLink to="/faq/protocol#wallet-not-supported">FAQ</HashLink>.
               </p>
             </>
           }
